Working on image_resize_interpolate, but this routine is pretty broken.
git-svn-id: svn://svn.code.sf.net/p/jp2a/code@480 a082123e-ac16-0410-b537-addfd1976b9b

+++ include/options.h
@@ -28,7 +28,7 @@ extern int debug;
 extern int clearscr;
 extern char ascii_palette[];
 extern float redweight, greenweight, blueweight;
-extern float RED[256], GREEN[256], BLUE[256], GRAY[256];
+extern unsigned short int RED[256], GREEN[256], BLUE[256], GRAY[256];
 extern const char *fileout;
 extern int usecolors;
 extern int termfit;
